requirements.txt
azure-functions
pandas>=1.0.3
matplotlib
scipy
numpy
